requirements.txt
google-cloud-aiplatform
google-cloud-firestore
flask
python-dotenv
